proyecto_comida_rapida/requirements.txt
Flask==2.3.3
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3 